# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VT Tracker API",
        "description": "API documentation for VT Tracker Application",
        "contact": {
            "email": "admin@vttracker.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/units/{unitId}/ride-comfort": {
            "get": {
                "tags": [
                    "Ride Comfort"
                ],
                "summary": "List Ride Comfort Results for Unit",
                "operationId": "c15208b3ce753d519227d042f4368a2e",
                "parameters": [
                    {
                        "name": "unitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RideComfortResult"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ride Comfort"
                ],
                "summary": "Submit Ride Comfort Result",
                "operationId": "0aa51ecbc292e1700eff1972bf69951c",
                "parameters": [
                    {
                        "name": "unitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vibration_value",
                                    "noise_db",
                                    "jerk_value"
                                ],
                                "properties": {
                                    "vibration_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "noise_db": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "jerk_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "device_used": {
                                        "type": "string",
                                        "enum": [
                                            "eva_625",
                                            "vibxpert_ii",
                                            "lms_test_lab",
                                            "kone_ride_check",
                                            "bruel_kjaer_2250",
                                            "other_certified"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Result Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RideComfortResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/units/{unitId}/stages": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "List Stages for Unit",
                "operationId": "853207a456e4519da128575759f22633",
                "parameters": [
                    {
                        "name": "unitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of stages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UnitStage"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stages/{id}": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "Show Stage Details with Tasks",
                "operationId": "b142b1e896d866f898e485a74b771690",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stage Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/UnitStage"
                                        },
                                        {
                                            "properties": {
                                                "tasks": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UnitTask"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Stages"
                ],
                "summary": "Update Stage Status",
                "operationId": "621d38556134c86662d857d345205e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stage Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnitStage"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RideComfortResult": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "unit_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "vibration_value": {
                        "type": "number",
                        "format": "float"
                    },
                    "noise_db": {
                        "type": "number",
                        "format": "float"
                    },
                    "jerk_value": {
                        "type": "number",
                        "format": "float"
                    },
                    "passed": {
                        "type": "boolean"
                    },
                    "measured_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "device_used": {
                        "type": "string",
                        "enum": [
                            "eva_625",
                            "vibxpert_ii",
                            "lms_test_lab",
                            "kone_ride_check",
                            "bruel_kjaer_2250",
                            "other_certified"
                        ],
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StageTemplate": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "stage_number": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "TaskTemplate": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "task_code": {
                        "type": "string"
                    },
                    "order_index": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UnitStage": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "unit_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "stage_template_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed"
                        ]
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "template": {
                        "$ref": "#/components/schemas/StageTemplate"
                    }
                },
                "type": "object"
            },
            "UnitTask": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "unit_stage_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "task_template_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "pass",
                            "fail"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "checked_by": {
                        "type": "integer",
                        "nullable": true
                    },
                    "checked_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "template": {
                        "$ref": "#/components/schemas/TaskTemplate"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}